+++ changelogs/en/CHANGELOG-1.3.x.md
@@ -255,6 +255,7 @@ Bug fixes, improvements on PDF export function, added major support for the webp
 - [x] Added better spacing for the image export of PDF, preventing images to become crowd in many cases 
 - [x] Added batch markdown export command `Sync repository to the website notes folder` to sync the entire repository to the target website notes folder.
 - [x] Added `sync files/folders to the website notes folder` options to the file explorer, Added `sync current file to website notes folder` command.
+- [ ] Added an option in settingsTab for the rendering of image captions/descriptions 
 - [ ] **Added support for Chinese translation of the SettingsTab Options and Commands** 
 
 📖 Documetations : 
@@ -273,13 +274,15 @@ Bug fixes, improvements on PDF export function, added major support for the webp
 - [x] fixed [#164](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/164), code blocks now render correctly in callouts. 
 - [x] fixed critical [#173](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/173), the PDF export can work correctly at the case some disturbing dollars are put inline. 
 - [x] fixed [#172](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/172), when export without meta-data, the image width are still kept. 
-- [ ] fixed [#175](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/175)
+- [x] fixed [#175](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/175), add recognition for native svg element of excalidraw when adding the caption and description for figures
+
 
 🏗️ refactors : 
 
 - [x] refactored and shortened pdf export code part
 - [x] refactored `getEquationAtCursor` and function in box current equations 
 - [x] refactored figure citation wrapper function 
+- [ ] refactored the figure caption creation extension part for better mainability 
 - [ ] resolved sonar issues to make sure that the code quality gate is passed. 
 
 ### 🚢 If you want following features, you can request it in issue page :
