+++ pnpm-workspace.yaml
@@ -3,3 +3,5 @@ allowBuilds:
   esbuild: set this to true or false
   obsidian-typings: set this to true or false
   unrs-resolver: set this to true or false
+minimumReleaseAgeExclude:
+  - '@friedparrot/equation-citator@1.3.7'
